refactor: move AiTutorContainer from CodeWorkspaceContainer to InstructionsWithWorkspace
