+++ apps/src/aiTutor/views/legacyLabs/AiTutorSidebar.module.scss
@@ -2,6 +2,7 @@
   width: 45px;
   position: absolute;
   right: 0;
+  top: 0;
 
   &-header {
     display: flex;
