+++ apps/src/templates/CodeWorkspaceContainer.jsx
@@ -9,9 +9,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import {connect} from 'react-redux';
 
-import {AiTutorContainer} from '../aiTutor/views/legacyLabs/AiTutorContainer';
 import commonStyles from '../commonStyles';
-import experiments from '../util/experiments';
 import * as utils from '../utils';
 
 class CodeWorkspaceContainer extends React.Component {
@@ -23,11 +21,6 @@ class CodeWorkspaceContainer extends React.Component {
     hidden: PropTypes.bool.isRequired,
     isRtl: PropTypes.bool.isRequired,
     noVisualization: PropTypes.bool.isRequired,
-    labType: PropTypes.string,
-  };
-
-  state = {
-    aiChatOpen: false,
   };
 
   /**
@@ -44,18 +37,8 @@ class CodeWorkspaceContainer extends React.Component {
     }
   }
 
-  toggleAiChat = () => {
-    this.setState(prevState => ({aiChatOpen: !prevState.aiChatOpen}));
-  };
-
   render() {
-    const {hidden, isRtl, noVisualization, children, style, labType} =
-      this.props;
-
-    const AiTutorLabs = ['applab', 'gamelab', 'weblab'];
-    const showAiTutor =
-      AiTutorLabs.includes(labType) &&
-      experiments.isEnabled(experiments.LEGACY_LAB_AI_TUTOR);
+    const {hidden, isRtl, noVisualization, children, style} = this.props;
 
     const mainStyle = {
       ...styles.main,
@@ -68,22 +51,9 @@ class CodeWorkspaceContainer extends React.Component {
 
     return (
       <div style={mainStyle} className="editor-column">
-        <div
-          id="codeWorkspace"
-          style={{
-            ...styles.codeWorkspace,
-            // 45px sidebar + 6px border = 51px
-            right: showAiTutor ? (this.state.aiChatOpen ? 350 : 51) : 0,
-          }}
-        >
+        <div id="codeWorkspace" style={styles.codeWorkspace}>
           {children}
         </div>
-        {showAiTutor && (
-          <AiTutorContainer
-            toggleAiChat={this.toggleAiChat}
-            aiChatOpen={this.state.aiChatOpen}
-          />
-        )}
       </div>
     );
   }
@@ -97,7 +67,6 @@ export default connect(
       !state.pageConstants.visualizationInWorkspace,
     isRtl: state.isRtl,
     noVisualization: state.pageConstants.noVisualization,
-    labType: state.pageConstants.appType,
   }),
   undefined,
   null,
